% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{create_dates_dataframe}
\alias{create_dates_dataframe}
\title{Create a data.frame with fields derived from the dates specified.}
\usage{
create_dates_dataframe(dates)
}
\arguments{
\item{dates}{Dates (character)}
}
\value{
data.frame with date related fields (date, year, month, day, hydro_year, season, hydr_season, apr1okt1 )
}
\description{
The newly created fiels can be used in selections
}
\examples{
dates <- as.Date( c("2009-01-02", "2009-01-03", "2009-01-04", "2009-01-05", "2009-01-06") )
create_dates_dataframe( dates )
}
