% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileutils.r
\docType{package}
\name{fileutils}
\alias{fileutils}
\title{fileutils: Collection of functions to handle and maniputate filenames.}
\description{
This package does not export any objects.
}
\details{
This package exports the following functions:
\itemize{
\item \code{\link{get_filename_extension}}
\item \code{\link{change_filename_extension}}
\item \code{\link{bare_filename}}
\item \code{\link{pks_dependencies}}
\item \code{\link{get_relative_filename}}
\item \code{\link{get_inp_filename}}
}
}
