% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{filter_idfnames}
\alias{filter_idfnames}
\title{Filter filenames of idf- files with a date included, like 'HEAD_20080402_l1.idf'}
\usage{
filter_idfnames(idfnames, fltr = NULL)
}
\arguments{
\item{idfnames}{idf-filename (character vector)}

\item{fltr}{filter to be applied, where (character)}
}
\value{
Selected names of idf-files.
}
\description{
Filter filenames of idf- files with a date included, like 'HEAD_20080402_l1.idf'
}
\examples{
idfnames <- c("HEAD_20080401_l1.idf","HEAD_20080501_l1")
fltr <- "month(dates)==4"
filter_idfnames(idfnames, fltr)
}
