% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{get_inp_filename}
\alias{get_inp_filename}
\title{Determine a script’s filename from within the script itself to create a new filename with a specified extension.}
\usage{
get_inp_filename(new_ext = ".inp")
}
\arguments{
\item{new_ext}{new extension (with leading dot) (character)}
}
\description{
Determine a script’s filename from within the script itself to create a new filename with a specified extension.
}
