% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{get_relative_filename}
\alias{get_relative_filename}
\title{Construct the pathname/filename with new basepath.}
\usage{
get_relative_filename(
  a_filename,
  relativeTo,
  new_basepath,
  new_ext = NULL,
  create_dir = FALSE
)
}
\arguments{
\item{a_filename}{filename (character).}

\item{relativeTo}{Reference pathname (character)}

\item{new_basepath}{new base path (character)}

\item{new_ext}{new extension (with leading dot) (character)}

\item{create_dir}{Create folder (boolean)}
}
\description{
Construct the pathname/filename with new basepath.
}
\examples{
a_filename <- "c:/input/input_sub/test.txt"
relativeTo <- "c:/input"
new_basepath <- "d:/results"
new_ext <- ".tif"
create_dir <- FALSE
get_relative_filename( a_filename, relativeTo, new_basepath, new_ext, create_dir )
}
