% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{rSIF_get_label_and_exprstr}
\alias{rSIF_get_label_and_exprstr}
\title{Return the label and expression part of a string created in DOS-batch file as an environment variable.}
\usage{
rSIF_get_label_and_exprstr(x)
}
\arguments{
\item{x}{String defined in a batch file and imported with 'Sys.getenv() (character)}
}
\value{
list with two character strings: 'label' and 'exprstr'. 'Label' characterises the dplyr 'summarise' result.
}
\description{
Background: All DOS versions interpret certain characters before executing a command.
Some well know examples are the percent sign ( \% ), and the redirection symbols ( < | > ).
These characters have to be escaped or surrounded by double quotes in order to pass them to an R-script as (part of) an system environment variable.
}
\details{
This function splits the string defined in a batch file imported with 'Sys.getenv() in two parts and returns a named list with two character strings: 'label' and 'exprstr'.
The result 'exprstr' can be used in a dplyr 'filter' function.
}
\examples{
x <- "sum_apr1okt1=group_by(apr1okt1)  summarise_all(.funs=c(gem='mean'))"
rSIF_get_label_and_exprstr(x)
}
