% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{rSIF_repair_exprstr_from_batch}
\alias{rSIF_repair_exprstr_from_batch}
\title{Repair an expression string that is imported from a DOS-batch file as an environment variable.}
\usage{
rSIF_repair_exprstr_from_batch(x)
}
\arguments{
\item{x}{Imported string from a DOS-batch file (character)}
}
\value{
Repaired expression string (character)
}
\description{
Remark: Double spaces are interpreted as \%>\%
}
\examples{
x <- "filter( apr1okt1 \"=\" \"other\")  group_by(apr1okt1)  summarise_all(.funs=c(gem='mean'))"
rSIF_repair_exprstr_from_batch(x)
}
