% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{splitstr_in_label_and_expression}
\alias{splitstr_in_label_and_expression}
\title{Split string in label and expression where '=' is the separator character.}
\usage{
splitstr_in_label_and_expression(x)
}
\arguments{
\item{x}{String with label and expression (character)}
}
\value{
list with two items: 'label' (character) and 'exprstr' (character)
}
\description{
Split string in label and expression where '=' is the separator character.
}
\examples{
x <- "sum_by_season= group_by(season) \%>\% summarise_all( .funs = c(gemiddelde='mean', sd='sd'))"
splitstr_in_label_and_expression(x)
}
